pyproject.toml
[tool.poetry]
name = "drawing"
version = "0.1.0"
description = "Research into training a model to draw outlines from photos with edge detection"
authors = ["Juliette <bourquinjuliette@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
numpy = "^1.19.4"
opencv-contrib-python = "^4.4.0"
typer = "^0.3.2"
pillow = "^8.0.1"
matplotlib = "^3.3.3"
tqdm = "^4.54.1"
spacy = "^2.3.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
